Move all functions in link.rs copied from cg_llvm to link_copied.rs
